+++ wok_code/wok_code/do_git_checkout_new_branch.py
@@ -9,7 +9,7 @@ import argparse
 
 from z0lib import z0lib
 
-__version__ = "2.0.18"
+__version__ = "2.0.19"
 
 
 class RepoCheckout(object):
@@ -422,6 +422,7 @@ def main(cli_args=None):
 
 if __name__ == "__main__":
     exit(main(None))
+