+++ wok_code/wok_code/scripts/cvt_csv_2_rst.py
@@ -39,7 +39,7 @@ except ImportError:
     import z0lib
 
 
-__version__ = "2.0.18"
+__version__ = "2.0.19"
 
 DOUBLE_CHARS = _u("✅❌")
 
@@ -211,6 +211,7 @@ def main(cli_args=None):
 
 if __name__ == "__main__":
     exit(main())
+