+++ wok_code/wok_code/scripts/please_python.py
@@ -10,7 +10,7 @@ else:
 
 from z0lib import z0lib
 
-__version__ = "2.0.18"
+__version__ = "2.0.19"
 
 
 class PleasePython(object):
@@ -123,6 +123,7 @@ BUGS
             )
             return 0
         return please.run_traced(cmd)
+