+++ wok_code/wok_code/tests/test_odoo_translation.py
@@ -11,7 +11,7 @@ import sys
 from z0lib import z0lib
 from zerobug import z0test, z0testodoo
 
-__version__ = "2.0.17"
+__version__ = "2.0.18"
 
 TEST_FAILED = 1
 TEST_SUCCESS = 0